fix: address review findings
- reject bodies missing data.types with 400 (was 500 via getHash TypeError)
- move messageRequestSchema to src/validation.ts so drizzle-kit and db.ts
  no longer load zod/@ethersproject just to read table definitions
- pin zod to ^3.25.76 and sync lockfile
- extract MAX_TIMESTAMP constant, document load-bearing refine order
