+++ package.json
@@ -32,7 +32,7 @@
     "semver": "^7.3.5",
     "ts-node": "^10.8.1",
     "typescript": "^5.9.3",
-    "zod": "^3"
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@snapshot-labs/eslint-config": "^0.1.1",
