+++ src/schema.ts
@@ -1,4 +1,3 @@
-import { getAddress } from '@ethersproject/address';
 import {
   bigint,
   index,
@@ -7,7 +6,6 @@ import {
   text,
   varchar
 } from 'drizzle-orm/pg-core';
-import { z } from 'zod';
 
 export const messages = pgTable(
   'messages',
@@ -29,33 +27,3 @@ export const messages = pgTable(
     index('messages_msg_hash_idx').on(table.msg_hash)
   ]
 );
-
-const isChecksummed = (a: unknown) => {
-  try {
-    return getAddress(a as string) === a;
-  } catch {
-    return false;
-  }
-};
-
-export const messageRequestSchema = z
-  .any()
-  .refine(b => b?.data?.message, { message: 'Invalid format request' })
-  .refine(
-    b =>
-      !b?.data?.message ||
-      b.data.types?.Space ||
-      b.data.message.settings ||
-      b.data.message.space,
-    { message: 'Missing space' }
-  )
-  .refine(b => isChecksummed(b?.address), { message: 'Invalid address' })
-  .refine(
-    b => {
-      if (!b?.data?.message) return true;
-      const ts = b.data.message.timestamp;
-      // Upper bound rejects ms-epoch mistakes and bigint-overflow values
-      return Number.isInteger(ts) && ts > 0 && ts <= 10_000_000_000;
-    },
-    { message: 'Invalid timestamp' }
-  );
